requirements.txt
fastapi
uvicorn
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-zipkin
python-jose
passlib
python-multipart
bcrypt
pyjwt
aiohttp
python-dotenv
PyPDF2
openai>=1.0.0
faiss-cpu
numpy
aiohttp
python-multipart
async-timeout